State Equals method moved from statemanager to state
