+++ packages/chain/statemanager/state_manager_test.go
@@ -172,7 +172,7 @@ func TestCruelWorld(t *testing.T) { //nolint:gocyclo
 				newBlockIndex+1, oldBlockIndex+1, peeringURLs[nodeIndex], err)
 			return false
 		}
-		if !sm_gpa_utils.StatesEqual(expectedNewState, results.GetNewState()) {
+		if !expectedNewState.Equals(results.GetNewState()) {
 			t.Logf("Mempool state request for new block %v and old block %v to node %v return wrong new state: expected trie root %s, received %s",
 				newBlockIndex+1, oldBlockIndex+1, peeringURLs[nodeIndex], blocks[newBlockIndex].TrieRoot(), results.GetNewState().TrieRoot())
 			return false
